% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{date_info}
\alias{date_info}
\title{\code{date_info}}
\format{
An object of class \code{date_info} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 9915 rows and 13 columns.
}
\usage{
date_info
}
\description{
\code{date_info}
}
\keyword{datasets}
