% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-db.R
\name{dump_all}
\alias{dump_all}
\title{Dump all the data linked to entities}
\usage{
dump_all(conn, ID_SITE, ID_ENTITY, entity_name, quiet = TRUE)
}
\arguments{
\item{conn}{DB connection object.}

\item{ID_SITE}{Optional, if \code{ID_ENTITY} or \code{entity_name} are provided.}

\item{ID_ENTITY}{Optional, if \code{ID_SITE} or \code{entity_name} are provided.}

\item{entity_name}{Optional, if \code{ID_SITE} or \code{ID_ENTITY} are provided.}

\item{quiet}{Boolean flag to indicate if queries should be displayed.}
}
\value{
List with the individual tables.
}
\description{
Dump all the data linked to entities
}
