% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{EPD_COUNTS}
\alias{EPD_COUNTS}
\title{\code{EPD_COUNTS}}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 99912 rows and 4842 columns.
}
\usage{
EPD_COUNTS
}
\description{
\code{EPD_COUNTS}
}
\keyword{datasets}
