% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{EPD_DATES}
\alias{EPD_DATES}
\title{\code{EPD_DATES}}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 9622 rows and 16 columns.
}
\usage{
EPD_DATES
}
\description{
\code{EPD_DATES}
}
\keyword{datasets}
