% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{EPD_DATES_coretops}
\alias{EPD_DATES_coretops}
\title{\code{EPD_DATES_coretops}}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 738 rows and 13 columns.
}
\usage{
EPD_DATES_coretops
}
\description{
\code{EPD_DATES_coretops}
}
\keyword{datasets}
