% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{EPD_METADATA}
\alias{EPD_METADATA}
\title{\code{EPD_METADATA}}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 1676 rows and 16 columns.
}
\usage{
EPD_METADATA
}
\description{
\code{EPD_METADATA}
}
\keyword{datasets}
