% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{EPD_taxa_amalgamation}
\alias{EPD_taxa_amalgamation}
\title{\code{EPD_taxa_amalgamation}}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 3300 rows and 7 columns.
}
\usage{
EPD_taxa_amalgamation
}
\description{
\code{EPD_taxa_amalgamation}
}
\keyword{datasets}
