% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{EPD_taxa_to_exclude}
\alias{EPD_taxa_to_exclude}
\title{\code{EPD_taxa_to_exclude}}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 1395 rows and 7 columns.
}
\usage{
EPD_taxa_to_exclude
}
\description{
\code{EPD_taxa_to_exclude}
}
\keyword{datasets}
