% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-entity.R
\name{get_entity}
\alias{get_entity}
\title{Get entities from the entity table}
\usage{
get_entity(conn, ID_ENTITY)
}
\arguments{
\item{conn}{DB connection object.}

\item{ID_ENTITY}{Numeric vector with IDs linked to entities.}
}
\value{
Tibble with entities and external links.
}
\description{
Get entities from the entity table
}
