% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{IBERIA_pollen_dates_v2}
\alias{IBERIA_pollen_dates_v2}
\title{\code{IBERIA_pollen_dates_v2}}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 950 rows and 15 columns.
}
\usage{
IBERIA_pollen_dates_v2
}
\description{
\code{IBERIA_pollen_dates_v2}
}
\keyword{datasets}
