% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{IBERIA_pollen_v3}
\alias{IBERIA_pollen_v3}
\title{\code{IBERIA_pollen_v3}}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 8779 rows and 222 columns.
}
\usage{
IBERIA_pollen_v3
}
\description{
\code{IBERIA_pollen_v3}
}
\keyword{datasets}
