% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{model_name}
\alias{model_name}
\title{\code{model_name}}
\format{
An object of class \code{model_name} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 18 rows and 2 columns.
}
\usage{
model_name
}
\description{
\code{model_name}
}
\keyword{datasets}
