% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sample}
\alias{sample}
\title{Pollen samples}
\format{
A data frame (\code{tibble} object) with
103091\verb{ rows and 11} variables:
\describe{
\item{ID_ENTITY}{\code{Unsigned integer}: Unique identifier for the entity
(as in the \code{entity} table)}
\item{ID_SAMPLE}{\code{Unsigned integer}: Unique identifier for each
sample}
\item{depth}{\code{Text}: Average sampling depth, in centimetres}
\item{thickness}{\code{Text}: Sample thickness, in centimetres}
\item{chronology_name}{\code{Text}: Name assigned to the chronology}
\item{age_type}{\code{Text}: Type of dating for the sample}
\item{age}{\code{Integer}: Value dated for the sample}
\item{age_younger}{\code{Integer}: Lower bound for the sample's age}
\item{age_older}{\code{Integer}: Upper bound for the sample's age}
\item{count_type}{\code{Text}: Type of count for the sample}
\item{sample_type}{\code{Text}: Type of sample}
}
}
\usage{
sample
}
\description{
Contains metadata for each pollen sample record in the database, \code{sample}
table in the database.
}
\examples{
sample_tb <- special.epd::sample
head(sample_tb)
}
\keyword{datasets}
