% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-db.R
\name{snapshot}
\alias{snapshot}
\alias{snapshot.MariaDBConnection}
\alias{snapshot.character}
\alias{snapshot.tbl_df}
\alias{snapshot.tbl}
\alias{snapshot.data.frame}
\alias{snapshot.default}
\title{Create a snapshot of the data linked to entities}
\usage{
snapshot(x, ...)

\method{snapshot}{MariaDBConnection}(x, ID_SITE, ID_ENTITY, entity_name, quiet = TRUE)

\method{snapshot}{character}(x, ...)

\method{snapshot}{tbl_df}(x, ...)

\method{snapshot}{tbl}(x, ...)

\method{snapshot}{data.frame}(x, ...)

\method{snapshot}{default}(x, ID_SITE, ID_ENTITY, entity_name)
}
\arguments{
\item{x}{This object accepts different classes. If the given object is
a database connection, then extracts data from the database using the
\code{ID_SITE}, \code{ID_ENTITY} or \code{entity_name} (these should be provided after
the connection object). Alternatively, if the given object is a vector,
then it will retrieve the records from an internal snapshot of the
database, included in this package.}

\item{...}{Optional parameters.}

\item{ID_SITE}{Optional, if \code{ID_ENTITY} or \code{entity_name} are provided.}

\item{ID_ENTITY}{Optional, if \code{ID_SITE} or \code{entity_name} are provided.}

\item{entity_name}{Optional, if \code{ID_SITE} or \code{ID_ENTITY} are provided.}

\item{quiet}{Boolean flag to indicate if queries should be displayed.}
}
\value{
List with the individual tables.
}
\description{
Create a snapshot of the data linked to entities. Including metadata,
dates, samples, age models and pollen counts.
}
