% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{package}
\name{special.epd-package}
\alias{special.epd}
\alias{special.epd-package}
\title{special.epd: SPECIAL Research Group's version of the European Pollen Database (EPD)}
\description{
This is a data package to provide access to records from the European Pollen Database (EPD), that have been checked by members of the SPECIAL reseach group. We have added age models using `rbacon` for each entity/core.
}
\author{
\strong{Maintainer}: Roberto Villegas-Diaz \email{r.villegas-diaz@reading.ac.uk} (\href{https://orcid.org/0000-0001-5036-8661}{ORCID})

Other contributors:
\itemize{
  \item SPECIAL Research Group @ University of Reading [copyright holder]
}

}
\keyword{internal}
