% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{taxon_name}
\alias{taxon_name}
\title{Taxon names}
\format{
A data frame (\code{tibble} object) with
2940\verb{ rows and 2} variables:
\describe{
\item{ID_TAXON}{\code{Unsigned integer}: Unique identifier for the taxon
name}
\item{taxon_name}{\code{Text}: }
}
}
\usage{
taxon_name
}
\description{
Contains a list of valid taxa. Represents the \code{taxon_name} table
in the database.
}
\examples{
taxon_name_tb <- special.epd::taxon_name
head(taxon_name_tb)
}
\keyword{datasets}
