% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{age_model}
\alias{age_model}
\title{\code{age_model}}
\format{
An object of class \code{age_model} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 89213 rows and 8 columns.
}
\usage{
age_model
}
\description{
\code{age_model}
}
\keyword{datasets}
