% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{external_link}
\alias{external_link}
\title{\code{external_link}}
\format{
An object of class \code{external_link} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 1978 rows and 7 columns.
}
\usage{
external_link
}
\description{
\code{external_link}
}
\keyword{datasets}
