% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sample}
\alias{sample}
\title{\code{sample}}
\format{
An object of class \code{sample} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 103091 rows and 11 columns.
}
\usage{
sample
}
\description{
\code{sample}
}
\keyword{datasets}
