% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-db.R
\name{snapshot}
\alias{snapshot}
\alias{snapshot.MariaDBConnection}
\alias{snapshot.default}
\title{Create a snapshot of the data linked to entities}
\usage{
snapshot(conn, ...)

\method{snapshot}{MariaDBConnection}(conn, ID_SITE, ID_ENTITY, entity_name, quiet = TRUE)

\method{snapshot}{default}(conn, ID_SITE, ID_ENTITY, entity_name)
}
\arguments{
\item{conn}{DB connection object. If none (or invalid connection object) is
passed, then defaults to use the latest version of the data, included
in this package.}

\item{...}{Optional parameters.}

\item{ID_SITE}{Optional, if \code{ID_ENTITY} or \code{entity_name} are provided.}

\item{ID_ENTITY}{Optional, if \code{ID_SITE} or \code{entity_name} are provided.}

\item{entity_name}{Optional, if \code{ID_SITE} or \code{ID_ENTITY} are provided.}

\item{quiet}{Boolean flag to indicate if queries should be displayed.}
}
\value{
List with the individual tables.

List with the individual tables.
}
\description{
Create a snapshot of the data linked to entities. Including metadata,
dates, samples, age models and pollen counts.
}
