% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-db.R
\name{write_csvs}
\alias{write_csvs}
\title{Write DB snapshot to disk}
\usage{
write_csvs(.data, prefix)
}
\arguments{
\item{.data}{DB snapshot (object of \code{special.epd} class).}

\item{prefix}{String with a prefix path where the data should be stored.}
}
\value{
Invisibly returns the input DB snapshot.
}
\description{
Write DB snapshot to disk as individual CSV files.
}
